+++ libMx/mx_camac.h
@@ -67,26 +67,26 @@ typedef struct {
  */
 
 typedef struct {
-	mx_status_type ( *get_lam_status ) ( MX_CAMAC *crate, int *lam_n );
-	mx_status_type ( *controller_command ) ( MX_CAMAC *crate, int command );
-	mx_status_type ( *camac ) ( MX_CAMAC *crate, int slot, int subaddress,
-		int function_code, int32_t *data, int *Q, int *X );
+	mx_status_type ( *get_lam_status ) ( MX_CAMAC *crate, long *lam_n );
+	mx_status_type ( *controller_command ) ( MX_CAMAC *crate, long command);
+	mx_status_type ( *camac ) ( MX_CAMAC *crate, long slot, long subaddress,
+		long function_code, int32_t *data, int *Q, int *X );
 } MX_CAMAC_FUNCTION_LIST;
 
 /* ============== Interface function prototypes. ============== */
 
 MX_API mx_status_type mx_camac_get_lam_status( MX_RECORD *camac_record,
-						int *lam_status );
+						long *lam_status );
 
 MX_API mx_status_type mx_camac_controller_command( MX_RECORD *camac_record,
-						int command );
+						long command );
 
 MX_API mx_status_type mx_camac( MX_RECORD *camac_record,
-				int slot, int subaddress, int function_code,
+				long slot, long subaddress, long function_code,
 				int32_t *data, int *Q, int *X );
 
 MX_API void mx_camac_qwait( MX_RECORD *camac_record,
-				int slot, int subaddress, int function_code,
+				long slot, long subaddress, long function_code,
 				int32_t *data, int *X );
 
 extern MX_RECORD_FUNCTION_LIST mx_camac_record_function_list;
